+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.3.0 (Unreleased)
+  * switch to associated type family for `Weak` inside `Weaken` - `Strengthen`
+    now has `Weaken` as a superclass
+    * I'm fairly confident that things make more sense this way - we get to
+      remove an open type family, improve type inference, and prevent users from
+      writing potentially dangerous instances. For that, a bit of asymmetry is
+      welcome.
+  * better document generic derivers
+  * various refactoring
+
 ## 0.2.0 (2022-05-31)
 Initial Hackage release (dependency issues prevented uploading).